move WebsocketClient and Crypto to a seperate directory so kaleb can use them
